+++ example/lib/main.dart
@@ -4,11 +4,13 @@ import 'package:flutter_map_marker_popup_example/popup_option_controls.dart';
 void main() => runApp(MyApp());
 
 /// NOTE:
-/// There are two main files:
+/// There are three main files:
 ///  - popup_option_controls.dart: This is just for toggling the different
 ///    options.
 ///  - map_with_popups.dart: This is where the actual map is and where you
 ///    can see an example of using the PopupMarkerLayerWidget.
+///  - simple_map_with_popups.dart: A simplified example of using the
+///    PopupMarkerLayerWidget.
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {